\name{mpg}
\alias{mpg}
\docType{data}
\title{Hocking(1976) Miles Per Gallon benchmark.}
\description{Performance data on 32 autos from Motor Trends magazine.}
\usage{data(mpg)}
\format{
  A data frame of 11 variables collected for 32 automobiles; no NAs.
  \describe{
    \item{cylnds}{number of cylinders.}
    \item{cubins}{cubic inches of engine displacement.}
    \item{hpower}{engine horsepower.}
    \item{weight}{total auto weight in pounds.}
    \item{mpg}{miles per gallon.}
    \item{shape}{engine shape (1=Straight, 0=V).}
    \item{transm}{transmission type.}
    \item{speeds}{number of forward speeds.}
    \item{carbs}{number of carburetors.}
    \item{fratio}{final drive rtaio.}
    \item{qmilt}{quarter mile time.}
  }
}
\references{
Hocking RA. "The Analysis and Selection of Variables in Regression." Biometrics 1976; 32: 1-51.
}
\keyword{datasets}